+++ src/cider/nrepl/middleware/test/extensions.clj
@@ -8,6 +8,39 @@
    [clojure.pprint :as pp]
    [clojure.test :as test :refer [assert-expr]]))
 
+(defn- diffable-objects?
+  "Is a diff between `a` and `b` worth displaying to the user?
+
+  Typically, two different scalar values don't have a useful diff.
+
+  Also, collections of different types are not meaningfully diffable."
+  [a b]
+  (and (or (and (map? a)
+                (map? b))
+
+           (and (sequential? a)
+                (sequential? b))
+
+           (and (set? a)
+                (set? b)))
+       (not= a b)))
+
+(defn maybe-assoc-diffs
+  "Computes and assocs data diffs when the at least expected/actual pair is deemed worth diffing."
+  [m expected actual]
+  (let [diffable? (volatile! false)
+        diffs (mapv (fn [actual-value]
+                      (when (diffable-objects? expected actual-value)
+                        (vreset! diffable? true))
+                      [actual-value (data/diff expected actual-value)])
+                    actual)]
+    (cond-> m
+      ;; If at least one actual value is diffable, then show diffs for all actual values,
+      ;; even if some of those diffs may not be so useful.
+      ;; Else one the client side could easily get confused.
+      @diffable?
+      (assoc :diffs diffs))))
+
 ;; From pjstadig/humane-test-output
 ;; https://github.com/pjstadig/humane-test-output
 (defn =-body
@@ -18,9 +51,9 @@
            result# (apply = expected# more#)]
        (->> (if result#
               {:type :pass}
-              {:type :fail
-               :diffs (->> (remove #(= expected# %) more#)
-                           (map #(vector % (data/diff expected# %))))})
+              (maybe-assoc-diffs {:type :fail}
+                                 expected#
+                                 (remove #(= expected# %) more#)))
             (merge {:message ~msg
                     :expected expected#
                     :actual